caNumeric extended to accept values errorless with prec of 10

+++ caQtDM_QtControls/src/enumeric.cpp
@@ -239,7 +239,7 @@ void ENumeric::init()
 
 void ENumeric::setValue(double v)
 {
-    long long temp = (long long) round(v * pow(10.0, decDig));
+    long long temp = (long long)round(v * (long long)pow(10.0, decDig));
     if ((temp >= minVal) && (temp <= maxVal)) {
         bool valChanged = data != temp;
         data = temp;
@@ -254,7 +254,7 @@ void ENumeric::setValue(double v)
 void ENumeric::silentSetValue(double v)
 {
     csValue = v;
-    long long temp = (long long) round(v * pow(10.0, decDig));
+    long long temp = (long long) round(v * (long long) pow(10.0, decDig));
     data = temp;
     showData();
 }
@@ -263,7 +263,7 @@ void ENumeric::setMaximum(double v)
 {
     if (v >= d_minAsDouble) {
         d_maxAsDouble = v;
-        maxVal = (long long) round(v* (long)pow(10.0, decDig));
+        maxVal = (long long) round(v* (long long)pow(10.0, decDig));
     }
 }
 
@@ -271,7 +271,7 @@ void ENumeric::setMinimum(double v)
 {
     if (v <= d_maxAsDouble) {
         d_minAsDouble = v;
-        minVal = (long long) round(v* (long)pow(10.0, decDig));
+        minVal = (long long) round(v* (long long)pow(10.0, decDig));
     }
 }
